Move *RestController.java and *Payload.java
*RestController went to cts/controller
*Payload went to cts/controller/payloads
This is meant to move us toward a DAO/model/service architecture
Unrelatedly, I also tried to fix the idtypes folder bug.
