+++ complete/src/main/java/org/theenergymashuplab/cts/GenerateClientCreateTender.java
@@ -17,6 +17,9 @@
 package org.theenergymashuplab.cts;
 
 import java.util.Random;
+
+import org.theenergymashuplab.cts.controller.payloads.ClientCreateTenderPayload;
+
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
