+++ complete/src/main/java/org/theenergymashuplab/cts/GreetingController.java
@@ -20,6 +20,11 @@ import java.util.concurrent.atomic.AtomicLong;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import org.theenergymashuplab.cts.controller.payloads.ClientCreateTenderPayload;
+import org.theenergymashuplab.cts.controller.payloads.ClientCreatedTenderPayload;
+import org.theenergymashuplab.cts.controller.payloads.EiCancelTenderPayload;
+import org.theenergymashuplab.cts.controller.payloads.EiCreateTenderPayload;
+import org.theenergymashuplab.cts.controller.payloads.EiCreateTransactionPayload;
 
 
 /*
