+++ complete/src/main/java/org/theenergymashuplab/cts/controller/LmaRestController.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.theenergymashuplab.cts;
+package org.theenergymashuplab.cts.controller;
 
 import java.util.concurrent.atomic.AtomicLong;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -31,7 +31,19 @@ import org.apache.logging.log4j.Logger;
 // For RestTemplate
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.web.client.RestTemplate;
+import org.theenergymashuplab.cts.controller.payloads.EICanceledTenderPayload;
+import org.theenergymashuplab.cts.controller.payloads.EiCancelTenderPayload;
+import org.theenergymashuplab.cts.controller.payloads.EiCreateTenderPayload;
+import org.theenergymashuplab.cts.controller.payloads.EiCreateTransactionPayload;
+import org.theenergymashuplab.cts.controller.payloads.EiCreatedTenderPayload;
+import org.theenergymashuplab.cts.controller.payloads.EiCreatedTransactionPayload;
 import org.theenergymashuplab.cts.controller.payloads.PositionAddPayload;
+import org.theenergymashuplab.cts.TenderIdType;
+import org.theenergymashuplab.cts.SideType;
+import org.theenergymashuplab.cts.Interval;
+import org.theenergymashuplab.cts.EiTransaction;
+import org.theenergymashuplab.cts.EiTender;
+import org.theenergymashuplab.cts.ActorIdType;
 
 @RestController
 @RequestMapping("/lma")
