+++ complete/src/main/java/org/theenergymashuplab/cts/controller/TeuaRestController.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.theenergymashuplab.cts;
+package org.theenergymashuplab.cts.controller;
 
 //import java.util.Random;
 //import java.util.concurrent.atomic.AtomicLong;
@@ -23,6 +23,21 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.client.RestTemplate;
+import org.theenergymashuplab.cts.ActorIdType;
+import org.theenergymashuplab.cts.EiResponse;
+import org.theenergymashuplab.cts.EiTender;
+import org.theenergymashuplab.cts.EiTransaction;
+import org.theenergymashuplab.cts.TenderIdType;
+import org.theenergymashuplab.cts.controller.payloads.ClientCreateTenderPayload;
+import org.theenergymashuplab.cts.controller.payloads.ClientCreateTransactionPayload;
+import org.theenergymashuplab.cts.controller.payloads.ClientCreatedTenderPayload;
+import org.theenergymashuplab.cts.controller.payloads.ClientCreatedTransactionPayload;
+import org.theenergymashuplab.cts.controller.payloads.EICanceledTenderPayload;
+import org.theenergymashuplab.cts.controller.payloads.EiCancelTenderPayload;
+import org.theenergymashuplab.cts.controller.payloads.EiCreateTenderPayload;
+import org.theenergymashuplab.cts.controller.payloads.EiCreateTransactionPayload;
+import org.theenergymashuplab.cts.controller.payloads.EiCreatedTenderPayload;
+import org.theenergymashuplab.cts.controller.payloads.EiCreatedTransactionPayload;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
