+++ complete/src/main/java/org/theenergymashuplab/cts/controller/payloads/EiCancelTenderPayload.java
@@ -14,7 +14,11 @@
  * limitations under the License.
  */
 
-package org.theenergymashuplab.cts;
+package org.theenergymashuplab.cts.controller.payloads;
+
+import org.theenergymashuplab.cts.ActorIdType;
+import org.theenergymashuplab.cts.RefIdType;
+import org.theenergymashuplab.cts.TenderIdType;
 
 public class EiCancelTenderPayload {
 	private ActorIdType counterPartyId;
