+++ complete/src/main/java/org/theenergymashuplab/cts/controller/payloads/EiCreatedTenderPayload.java
@@ -14,69 +14,62 @@
  * limitations under the License.
  */
 
-package org.theenergymashuplab.cts;
+package org.theenergymashuplab.cts.controller.payloads;
 
-public class EiCreatedTransactionPayload {
-	private TransactionIdType transactionId;
+import org.theenergymashuplab.cts.ActorIdType;
+import org.theenergymashuplab.cts.EiResponse;
+import org.theenergymashuplab.cts.RefIdType;
+import org.theenergymashuplab.cts.TenderIdType;
+
+public class EiCreatedTenderPayload {
+	private TenderIdType tenderId;
 	private ActorIdType partyId;
 	private ActorIdType counterPartyId;
 	public EiResponse response;
-//	public ArrayofResponses responses; NOT USED YET
+//	public ArrayofResponses responses; NOT USED
 	private final RefIdType refId = new RefIdType();
 
-	// Default initializer for JSON serialization
-	public EiCreatedTransactionPayload() {
+	/*
+	 * Default constructor for JSON deserialization.
+	 * TO DO change to zero Id values in ActorId and RefId constructors
+	 */
+	public EiCreatedTenderPayload()	{		
 	}
 	
-	public TransactionIdType getTransactionId() {
-		return transactionId;
-	}
-
-	public void setTransactionId(TransactionIdType transactionId) {
-		this.transactionId = transactionId;
-	}
-
-	public void setPartyId(ActorIdType partyId) {
-		this.partyId = partyId;
-	}
-
-	public void setCounterPartyId(ActorIdType counterPartyId) {
-		this.counterPartyId = counterPartyId;
-	}
-
-	public EiCreatedTransactionPayload(
-			TransactionIdType transactionId,
+	public EiCreatedTenderPayload(
+			TenderIdType tenderId,
 			ActorIdType partyId,
 			ActorIdType counterPartyId,
 			EiResponse response) {
 
-		this.transactionId = transactionId;
+		this.tenderId = tenderId;
 		this.partyId = partyId;
 		this.counterPartyId = counterPartyId;
 		this.response = response;
 	}
 
 	public long getId() {
-		return transactionId.value();
+		return tenderId.value();
 	}
 
 	public void print() {		
 		System.err.println(
-				"EiCreatedTransactionPayload transactionId " +
-				transactionId.toString() +
-				" partyId " + partyId.value() +
-				" counterPartyId " + counterPartyId.value() +
-				" refId " + refId.toString());
-	}
-	
-	public String toString()	{
-		return ("EiCreatedTransactionPayload transactionId " +
-				transactionId.toString() +
+				"EiCreatedTender tenderId " +
+				tenderId.toString() +
 				" partyId " + partyId.toString() +
 				" counterPartyId " + counterPartyId.toString() +
-				" refId " + refId.toString());
+				" refId " + refId.toString() +
+				" response " + response.toString());
 	}
 	
+	public String toString() {
+		return "EiCreatedTender tenderId " +
+				tenderId.toString() +
+				" partyId " + partyId.toString() +
+				" counterPartyId " + counterPartyId.toString() +
+				" refId " + refId.toString() +
+				" response " + response.toString();
+	}
 	
 	public EiResponse getResponse() {
 		return response;
@@ -84,6 +77,10 @@ public class EiCreatedTransactionPayload {
 
 	public void setResponse(EiResponse response) {
 		this.response = response;
+	}
+
+	public TenderIdType getTenderId() {
+		return tenderId;
 	}
 
 	public ActorIdType getPartyId() {
