+++ complete/src/main/java/org/theenergymashuplab/cts/controller/payloads/MarketCreatedTenderPayload.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.theenergymashuplab.cts;
+package org.theenergymashuplab.cts.controller.payloads;
 
 /*
  * Sent by Market responding to MarketCreateTender from LME.
